Creating find person by id

+++ src/main/java/com/github/myqandrade/personservice/repositories/AddressRepository.java
@@ -4,8 +4,6 @@ import com.github.myqandrade.personservice.model.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface AddressRepository extends JpaRepository<Address, Long> {
+public interface AddressRepository extends JpaRepository<Address, Integer> {
 }
